Fix bug where a page table shared across multiple VMMs would become de-synced
The bug here is that a running task can allocate memory within a shared page table, such as the kernel page tables that are shared between every process. The running task will mark the now-used memory as “used” within its bitmap allocator. The page tables will be changed in every process, but only the running process will have marked those pages as used in its allocator.
I added logic to test for updates to shared kernel tables on task-switch. For now, it panics when inconsistencies are detected, but I will build a mechanism to keep shared memory allocation states in-sync when it’s modified.