[UEFI] Mark paging structures as uncacheable
I set this while debugging an issue in which the LAPIC timer wouldn’t fire, but the issue ended up being that I was using register vectors that were multiplied by 0x10. I’m leaving this in because the APIC docs say that the memory should be mapped as uncacheable, but we should eventually change it so things are marked as uncacheable on-demand.